requirements.txt
time
pybase64
pywin32
pytest-shutil
pycryptodome
colorama
getpass4
datetime